KUX-1761: Upgrade exoplayer to Media 3
