+++ playkit/src/main/java/com/kaltura/androidx/media3/exoplayer/dashmanifestparser/CustomFormat.java
@@ -1,4 +1,4 @@
-package com.kaltura.android.exoplayer2.dashmanifestparser;
+package com.kaltura.androidx.media3.exoplayer.dashmanifestparser;
 
 import android.os.Bundle;
 import android.os.Parcel;
@@ -8,15 +8,15 @@ import androidx.annotation.IntDef;
 import androidx.annotation.Nullable;
 
 import com.google.common.base.Joiner;
-import com.kaltura.android.exoplayer2.Bundleable;
-import com.kaltura.android.exoplayer2.C;
-import com.kaltura.android.exoplayer2.Format;
-import com.kaltura.android.exoplayer2.drm.DrmInitData;
-import com.kaltura.android.exoplayer2.metadata.Metadata;
-import com.kaltura.android.exoplayer2.util.BundleableUtil;
-import com.kaltura.android.exoplayer2.util.MimeTypes;
-import com.kaltura.android.exoplayer2.util.Util;
-import com.kaltura.android.exoplayer2.video.ColorInfo;
+import com.kaltura.androidx.media3.common.Bundleable;
+import com.kaltura.androidx.media3.common.C;
+import com.kaltura.androidx.media3.common.Format;
+import com.kaltura.androidx.media3.common.DrmInitData;
+import com.kaltura.androidx.media3.common.Metadata;
+import com.kaltura.androidx.media3.common.util.BundleableUtil;
+import com.kaltura.androidx.media3.common.MimeTypes;
+import com.kaltura.androidx.media3.common.util.Util;
+import com.kaltura.androidx.media3.common.ColorInfo;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
