+++ playkit/src/main/java/com/kaltura/androidx/media3/exoplayer/video/KVideoRendererFirstFrameWhenStartedEventListener.java
@@ -1,4 +1,4 @@
-package com.kaltura.android.exoplayer2.video;
+package com.kaltura.androidx.media3.exoplayer.video;
 
 public interface KVideoRendererFirstFrameWhenStartedEventListener {
     default void onRenderedFirstFrameWhenStarted() {
