+++ playkit/src/main/java/com/kaltura/playkit/drm/DrmCallback.java
@@ -7,17 +7,17 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import com.google.common.collect.ImmutableMap;
-import com.kaltura.android.exoplayer2.C;
-import com.kaltura.android.exoplayer2.drm.ExoMediaDrm;
-import com.kaltura.android.exoplayer2.drm.HttpMediaDrmCallback;
-import com.kaltura.android.exoplayer2.drm.MediaDrmCallback;
-import com.kaltura.android.exoplayer2.drm.MediaDrmCallbackException;
-import com.kaltura.android.exoplayer2.upstream.DataSource;
-import com.kaltura.android.exoplayer2.upstream.DataSourceInputStream;
-import com.kaltura.android.exoplayer2.upstream.DataSpec;
-import com.kaltura.android.exoplayer2.upstream.HttpDataSource;
-import com.kaltura.android.exoplayer2.upstream.StatsDataSource;
-import com.kaltura.android.exoplayer2.util.Util;
+import com.kaltura.androidx.media3.common.C;
+import com.kaltura.androidx.media3.exoplayer.drm.ExoMediaDrm;
+import com.kaltura.androidx.media3.exoplayer.drm.HttpMediaDrmCallback;
+import com.kaltura.androidx.media3.exoplayer.drm.MediaDrmCallback;
+import com.kaltura.androidx.media3.exoplayer.drm.MediaDrmCallbackException;
+import com.kaltura.androidx.media3.datasource.DataSource;
+import com.kaltura.androidx.media3.datasource.DataSourceInputStream;
+import com.kaltura.androidx.media3.datasource.DataSpec;
+import com.kaltura.androidx.media3.datasource.HttpDataSource;
+import com.kaltura.androidx.media3.datasource.StatsDataSource;
+import com.kaltura.androidx.media3.common.util.Util;
 import com.kaltura.playkit.PKLog;
 import com.kaltura.playkit.PKRequestParams;
 import com.kaltura.playkit.player.MediaSupport;
