+++ playkit/src/main/java/com/kaltura/playkit/drm/SimpleDashParser.java
@@ -16,20 +16,20 @@ import android.net.Uri;
 import android.os.Build;
 import androidx.annotation.Nullable;
 
-import com.kaltura.android.exoplayer2.C;
-import com.kaltura.android.exoplayer2.Format;
-import com.kaltura.android.exoplayer2.drm.DrmInitData;
-import com.kaltura.android.exoplayer2.extractor.mp4.FragmentedMp4Extractor;
-import com.kaltura.android.exoplayer2.extractor.mp4.PsshAtomUtil;
-import com.kaltura.android.exoplayer2.source.chunk.BundledChunkExtractor;
-import com.kaltura.android.exoplayer2.source.chunk.InitializationChunk;
-import com.kaltura.android.exoplayer2.source.dash.manifest.AdaptationSet;
-import com.kaltura.android.exoplayer2.source.dash.manifest.DashManifest;
-import com.kaltura.android.exoplayer2.source.dash.manifest.DashManifestParser;
-import com.kaltura.android.exoplayer2.source.dash.manifest.Period;
-import com.kaltura.android.exoplayer2.source.dash.manifest.Representation;
-import com.kaltura.android.exoplayer2.upstream.DataSpec;
-import com.kaltura.android.exoplayer2.upstream.FileDataSource;
+import com.kaltura.androidx.media3.common.C;
+import com.kaltura.androidx.media3.common.Format;
+import com.kaltura.androidx.media3.common.DrmInitData;
+import com.kaltura.androidx.media3.extractor.mp4.FragmentedMp4Extractor;
+import com.kaltura.androidx.media3.extractor.mp4.PsshAtomUtil;
+import com.kaltura.androidx.media3.exoplayer.source.chunk.BundledChunkExtractor;
+import com.kaltura.androidx.media3.exoplayer.source.chunk.InitializationChunk;
+import com.kaltura.androidx.media3.exoplayer.dash.manifest.AdaptationSet;
+import com.kaltura.androidx.media3.exoplayer.dash.manifest.DashManifest;
+import com.kaltura.androidx.media3.exoplayer.dash.manifest.DashManifestParser;
+import com.kaltura.androidx.media3.exoplayer.dash.manifest.Period;
+import com.kaltura.androidx.media3.exoplayer.dash.manifest.Representation;
+import com.kaltura.androidx.media3.datasource.DataSpec;
+import com.kaltura.androidx.media3.datasource.FileDataSource;
 import com.kaltura.playkit.PKLog;
 import com.kaltura.playkit.player.MediaSupport;
 import com.kaltura.playkit.utils.Consts;
