+++ playkit/src/main/java/com/kaltura/playkit/player/BaseExoplayerView.java
@@ -3,9 +3,9 @@ package com.kaltura.playkit.player;
 import android.content.Context;
 import android.util.AttributeSet;
 
-import com.kaltura.android.exoplayer2.ExoPlayer;
-import com.kaltura.android.exoplayer2.ui.SubtitleView;
-import com.kaltura.android.exoplayer2.video.KVideoRendererFirstFrameWhenStartedEventListener;
+import com.kaltura.androidx.media3.exoplayer.ExoPlayer;
+import com.kaltura.androidx.media3.ui.SubtitleView;
+import com.kaltura.androidx.media3.exoplayer.video.KVideoRendererFirstFrameWhenStartedEventListener;
 
 public abstract class BaseExoplayerView extends PlayerView implements KVideoRendererFirstFrameWhenStartedEventListener {
