+++ playkit/src/main/java/com/kaltura/playkit/player/PKExternalSubtitle.java
@@ -4,8 +4,8 @@ import android.os.Parcel;
 import android.os.Parcelable;
 import androidx.annotation.NonNull;
 
-import com.kaltura.android.exoplayer2.C;
-import com.kaltura.android.exoplayer2.Format;
+import com.kaltura.androidx.media3.common.C;
+import com.kaltura.androidx.media3.common.Format;
 import com.kaltura.playkit.PKSubtitleFormat;
 
 public class PKExternalSubtitle implements Parcelable {
