+++ playkit/src/main/java/com/kaltura/playkit/player/TrackSelectionHelper.java
@@ -20,25 +20,25 @@ import android.util.Pair;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
-import com.kaltura.android.exoplayer2.C;
-import com.kaltura.android.exoplayer2.Format;
-import com.kaltura.android.exoplayer2.RendererCapabilities;
-import com.kaltura.android.exoplayer2.Tracks;
-import com.kaltura.android.exoplayer2.dashmanifestparser.CustomAdaptationSet;
-import com.kaltura.android.exoplayer2.dashmanifestparser.CustomDashManifest;
-import com.kaltura.android.exoplayer2.dashmanifestparser.CustomFormat;
-import com.kaltura.android.exoplayer2.dashmanifestparser.CustomRepresentation;
-import com.kaltura.android.exoplayer2.source.TrackGroup;
-import com.kaltura.android.exoplayer2.source.TrackGroupArray;
-import com.kaltura.android.exoplayer2.source.dash.manifest.EventStream;
-import com.kaltura.android.exoplayer2.text.Subtitle;
-import com.kaltura.android.exoplayer2.text.webvtt.WebvttCueInfo;
-import com.kaltura.android.exoplayer2.trackselection.DefaultTrackSelector;
-import com.kaltura.android.exoplayer2.trackselection.DefaultTrackSelector.SelectionOverride;
-import com.kaltura.android.exoplayer2.trackselection.MappingTrackSelector;
-import com.kaltura.android.exoplayer2.trackselection.TrackSelectionOverride;
-import com.kaltura.android.exoplayer2.trackselection.TrackSelectionParameters;
-import com.kaltura.android.exoplayer2.util.Util;
+import com.kaltura.androidx.media3.common.C;
+import com.kaltura.androidx.media3.common.Format;
+import com.kaltura.androidx.media3.exoplayer.RendererCapabilities;
+import com.kaltura.androidx.media3.common.Tracks;
+import com.kaltura.androidx.media3.exoplayer.dashmanifestparser.CustomAdaptationSet;
+import com.kaltura.androidx.media3.exoplayer.dashmanifestparser.CustomDashManifest;
+import com.kaltura.androidx.media3.exoplayer.dashmanifestparser.CustomFormat;
+import com.kaltura.androidx.media3.exoplayer.dashmanifestparser.CustomRepresentation;
+import com.kaltura.androidx.media3.common.TrackGroup;
+import com.kaltura.androidx.media3.exoplayer.source.TrackGroupArray;
+import com.kaltura.androidx.media3.exoplayer.dash.manifest.EventStream;
+import com.kaltura.androidx.media3.extractor.text.Subtitle;
+import com.kaltura.androidx.media3.extractor.text.webvtt.WebvttCueInfo;
+import com.kaltura.androidx.media3.exoplayer.trackselection.DefaultTrackSelector;
+import com.kaltura.androidx.media3.exoplayer.trackselection.DefaultTrackSelector.SelectionOverride;
+import com.kaltura.androidx.media3.exoplayer.trackselection.MappingTrackSelector;
+import com.kaltura.androidx.media3.common.TrackSelectionOverride;
+import com.kaltura.androidx.media3.common.TrackSelectionParameters;
+import com.kaltura.androidx.media3.common.util.Util;
 import com.kaltura.playkit.PKAbrFilter;
 import com.kaltura.playkit.PKAudioCodec;
 import com.kaltura.playkit.PKError;
@@ -65,11 +65,11 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 
-import static com.kaltura.android.exoplayer2.util.MimeTypes.AUDIO_AAC;
-import static com.kaltura.android.exoplayer2.util.MimeTypes.VIDEO_AV1;
-import static com.kaltura.android.exoplayer2.util.MimeTypes.VIDEO_H265;
-import static com.kaltura.android.exoplayer2.util.MimeTypes.VIDEO_VP8;
-import static com.kaltura.android.exoplayer2.util.MimeTypes.VIDEO_VP9;
+import static com.kaltura.androidx.media3.common.MimeTypes.AUDIO_AAC;
+import static com.kaltura.androidx.media3.common.MimeTypes.VIDEO_AV1;
+import static com.kaltura.androidx.media3.common.MimeTypes.VIDEO_H265;
+import static com.kaltura.androidx.media3.common.MimeTypes.VIDEO_VP8;
+import static com.kaltura.androidx.media3.common.MimeTypes.VIDEO_VP9;
 import static com.kaltura.playkit.utils.Consts.TRACK_TYPE_AUDIO;
 import static com.kaltura.playkit.utils.Consts.TRACK_TYPE_IMAGE;
 import static com.kaltura.playkit.utils.Consts.TRACK_TYPE_TEXT;
