+++ playkit/src/main/java/com/kaltura/playkit/profiler/ExoPlayerProfilingListener.java
@@ -1,28 +1,28 @@
 package com.kaltura.playkit.profiler;
 
 
-import static com.kaltura.android.exoplayer2.C.DATA_TYPE_AD;
-import static com.kaltura.android.exoplayer2.C.DATA_TYPE_DRM;
-import static com.kaltura.android.exoplayer2.C.DATA_TYPE_MANIFEST;
-import static com.kaltura.android.exoplayer2.C.DATA_TYPE_MEDIA;
-import static com.kaltura.android.exoplayer2.C.DATA_TYPE_MEDIA_INITIALIZATION;
-import static com.kaltura.android.exoplayer2.C.DATA_TYPE_TIME_SYNCHRONIZATION;
-import static com.kaltura.android.exoplayer2.C.DATA_TYPE_UNKNOWN;
-import static com.kaltura.android.exoplayer2.C.SELECTION_REASON_ADAPTIVE;
-import static com.kaltura.android.exoplayer2.C.SELECTION_REASON_INITIAL;
-import static com.kaltura.android.exoplayer2.C.SELECTION_REASON_MANUAL;
-import static com.kaltura.android.exoplayer2.C.SELECTION_REASON_TRICK_PLAY;
-import static com.kaltura.android.exoplayer2.C.SELECTION_REASON_UNKNOWN;
-import static com.kaltura.android.exoplayer2.C.TRACK_TYPE_AUDIO;
-import static com.kaltura.android.exoplayer2.C.TRACK_TYPE_DEFAULT;
-import static com.kaltura.android.exoplayer2.C.TRACK_TYPE_TEXT;
-import static com.kaltura.android.exoplayer2.C.TRACK_TYPE_VIDEO;
-import static com.kaltura.android.exoplayer2.Player.DISCONTINUITY_REASON_AUTO_TRANSITION;
-import static com.kaltura.android.exoplayer2.Player.DISCONTINUITY_REASON_INTERNAL;
-import static com.kaltura.android.exoplayer2.Player.DISCONTINUITY_REASON_REMOVE;
-import static com.kaltura.android.exoplayer2.Player.DISCONTINUITY_REASON_SEEK;
-import static com.kaltura.android.exoplayer2.Player.DISCONTINUITY_REASON_SEEK_ADJUSTMENT;
-import static com.kaltura.android.exoplayer2.Player.DISCONTINUITY_REASON_SKIP;
+import static com.kaltura.androidx.media3.common.C.DATA_TYPE_AD;
+import static com.kaltura.androidx.media3.common.C.DATA_TYPE_DRM;
+import static com.kaltura.androidx.media3.common.C.DATA_TYPE_MANIFEST;
+import static com.kaltura.androidx.media3.common.C.DATA_TYPE_MEDIA;
+import static com.kaltura.androidx.media3.common.C.DATA_TYPE_MEDIA_INITIALIZATION;
+import static com.kaltura.androidx.media3.common.C.DATA_TYPE_TIME_SYNCHRONIZATION;
+import static com.kaltura.androidx.media3.common.C.DATA_TYPE_UNKNOWN;
+import static com.kaltura.androidx.media3.common.C.SELECTION_REASON_ADAPTIVE;
+import static com.kaltura.androidx.media3.common.C.SELECTION_REASON_INITIAL;
+import static com.kaltura.androidx.media3.common.C.SELECTION_REASON_MANUAL;
+import static com.kaltura.androidx.media3.common.C.SELECTION_REASON_TRICK_PLAY;
+import static com.kaltura.androidx.media3.common.C.SELECTION_REASON_UNKNOWN;
+import static com.kaltura.androidx.media3.common.C.TRACK_TYPE_AUDIO;
+import static com.kaltura.androidx.media3.common.C.TRACK_TYPE_DEFAULT;
+import static com.kaltura.androidx.media3.common.C.TRACK_TYPE_TEXT;
+import static com.kaltura.androidx.media3.common.C.TRACK_TYPE_VIDEO;
+import static com.kaltura.androidx.media3.common.Player.DISCONTINUITY_REASON_AUTO_TRANSITION;
+import static com.kaltura.androidx.media3.common.Player.DISCONTINUITY_REASON_INTERNAL;
+import static com.kaltura.androidx.media3.common.Player.DISCONTINUITY_REASON_REMOVE;
+import static com.kaltura.androidx.media3.common.Player.DISCONTINUITY_REASON_SEEK;
+import static com.kaltura.androidx.media3.common.Player.DISCONTINUITY_REASON_SEEK_ADJUSTMENT;
+import static com.kaltura.androidx.media3.common.Player.DISCONTINUITY_REASON_SKIP;
 import static com.kaltura.playkit.profiler.PlayKitProfiler.MSEC_MULTIPLIER_FLOAT;
 import static com.kaltura.playkit.profiler.PlayKitProfiler.field;
 import static com.kaltura.playkit.profiler.PlayKitProfiler.joinFields;
@@ -37,20 +37,20 @@ import androidx.annotation.Nullable;
 import com.google.common.collect.ImmutableList;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
-import com.kaltura.android.exoplayer2.Format;
-import com.kaltura.android.exoplayer2.PlaybackException;
-import com.kaltura.android.exoplayer2.PlaybackParameters;
-import com.kaltura.android.exoplayer2.Player;
-import com.kaltura.android.exoplayer2.Tracks;
-import com.kaltura.android.exoplayer2.analytics.AnalyticsListener;
-import com.kaltura.android.exoplayer2.decoder.DecoderCounters;
-import com.kaltura.android.exoplayer2.decoder.DecoderReuseEvaluation;
-import com.kaltura.android.exoplayer2.drm.DrmSession;
-import com.kaltura.android.exoplayer2.metadata.Metadata;
-import com.kaltura.android.exoplayer2.source.LoadEventInfo;
-import com.kaltura.android.exoplayer2.source.MediaLoadData;
-import com.kaltura.android.exoplayer2.source.TrackGroup;
-import com.kaltura.android.exoplayer2.video.VideoSize;
+import com.kaltura.androidx.media3.common.Format;
+import com.kaltura.androidx.media3.common.PlaybackException;
+import com.kaltura.androidx.media3.common.PlaybackParameters;
+import com.kaltura.androidx.media3.common.Player;
+import com.kaltura.androidx.media3.common.Tracks;
+import com.kaltura.androidx.media3.exoplayer.analytics.AnalyticsListener;
+import com.kaltura.androidx.media3.exoplayer.DecoderCounters;
+import com.kaltura.androidx.media3.exoplayer.DecoderReuseEvaluation;
+import com.kaltura.androidx.media3.exoplayer.drm.DrmSession;
+import com.kaltura.androidx.media3.common.Metadata;
+import com.kaltura.androidx.media3.exoplayer.source.LoadEventInfo;
+import com.kaltura.androidx.media3.exoplayer.source.MediaLoadData;
+import com.kaltura.androidx.media3.common.TrackGroup;
+import com.kaltura.androidx.media3.common.VideoSize;
 import com.kaltura.playkit.PKPlaybackException;
 import com.kaltura.playkit.player.PKPlayerErrorType;
