+++ playkit/src/main/java/com/kaltura/playkit/profiler/PlayKitProfiler.java
@@ -17,8 +17,8 @@ import com.google.gson.Gson;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParseException;
-import com.kaltura.android.exoplayer2.C;
-import com.kaltura.android.exoplayer2.analytics.AnalyticsListener;
+import com.kaltura.androidx.media3.common.C;
+import com.kaltura.androidx.media3.exoplayer.analytics.AnalyticsListener;
 import com.kaltura.playkit.PKDrmParams;
 import com.kaltura.playkit.PKLog;
 import com.kaltura.playkit.PKMediaConfig;
